Add classes Ui, Storage, Parser, TaskList
