+++ data/duke.txt
@@ -1,7 +1,6 @@
 T|0|wak
 T|0|run
 T|0|fly
-T|0|run
-T|0|walk
-T|0|run
-D | 0 | return book | 2019-12-02T18:00
+T|1|run
+T|0|lit
+T|0|nbye
