+++ text-ui-test/input.txt
@@ -1,6 +1,6 @@
 todo borrow book
 list
-deadline return book /by Sunday
+deadline return book /by 2/12/2019 1800
 go for a walk
 event project meeting /from Mon 2pm /to 4pm
 list
